+++ CHANGELOG.md
@@ -18,6 +18,7 @@ API-breaking change is listed.
 - Add support for `negb` to `naive_solver` tactic. (by Janggun Lee)
 - Add lemma `map_size_alter`. (by Alban Reynaud Michez)
 - Add lemma `map_size_kmap`. (by Alban Reynaud Michez)
+- Add lemma `map_kmap_compose`. (by Alban Reynaud Michez)
 
 ## std++ 1.13.0 (2026-03-05)